Refactor (APIMF-3491): extracted external link property parser
